backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
dj-database-url==2.2.0
psycopg[binary]==3.2.1
python-dotenv==1.0.1
pytest==8.3.2
pytest-django==4.8.0
ruff==0.6.9
black==24.8.0
drf-spectacular==0.27.2
django-cors-headers==4.4.0